Redesigned the API for lava.utils.{slurm,loihi}.
